got the whole game working.
The game is pretty simple. It shows you a picture of a dog then you have to guess what it is. depending on your answer it will show you a gif based on wheather you got it right of not. It will keep track of your score. click next then keep going.